// src/app/models/reports/ORMrptEncounterCount.ts
export class ORMrptEncounterCount {
    location_name: String;
    provider_name: String;
    is_aggregate: Boolean;
    day_01: Number=0;
    day_02: Number=0;
    day_03: Number=0;
    day_04: Number=0;
    day_05: Number=0;
    day_06: Number=0;
    day_07: Number=0;
    day_08: Number=0;
    day_09: Number=0;
    day_10: Number=0;
    day_11: Number=0;
    day_12: Number=0;
    day_13: Number=0;
    day_14: Number=0;
    day_15: Number=0;
    day_16: Number=0;
    day_17: Number=0;
    day_18: Number=0;
    day_19: Number=0;
    day_20: Number=0;
    day_21: Number=0;
    day_22: Number=0;
    day_23: Number=0;
    day_24: Number=0;
    day_25: Number=0;
    day_26: Number=0;
    day_27: Number=0;
    day_28: Number=0;
    day_29: Number=0;
    day_30: Number=0;
    day_31: Number=0;
    total: Number=0;
}