// src/app/models/reports/ORMproviderEncounterCount.ts
export class ORMproviderEncounterCount{
    location_name: String;
    provider_name: String;
    is_aggregate: Boolean;
    total: Number=0;
    jan: Number=0;
    feb: Number=0;
    mar: Number=0;
    apr: Number=0;
    may: Number=0;
    jun: Number=0;
    jul: Number=0;
    aug: Number=0;
    sep: Number=0;
    oct: Number=0;
    nov: Number=0;
    dec: Number=0;
}